badger db initial implementation
