+++ .env.test
@@ -14,3 +14,6 @@ TRANSLATE_DB_MYSQL_PORT=3306
 TRANSLATE_DB_MYSQL_USER=root
 TRANSLATE_DB_MYSQL_PASSWORD=
 TRANSLATE_DB_MYSQL_DATABASE=translate
+
+# BadgerDB
+TRANSLATE_DB_BADGERDB_PATH=/tmp/badger
