+++ examples/translate.yaml
@@ -10,4 +10,5 @@ db:
     user: "root"
     password: ""
     database: "translate"
-  postgres:
+  badgerdb:
+    path: "/tmp/badgerdb"
